requirements.txt
streamlit
pandas
sentence-transformers
faiss-cpu
openai
python-dotenv